identifier: Add identifiers package; rename NewDNS & NewIP
Add a new `identifiers` package for functions that handle slices of `ACMEIdentifier`. We plan to add several such functions (e.g. in #7961). Without this change, they'd need a clumsy prefix/suffix to differentiate them from their sibling functions that handle single idents. The current `NewDNS` / `FromDNSNames` are the first example of this.

Rename `NewDNS` and `NewIP` to use a more standardized `From` prefix, since they're doing conversions. Move and rename `FromDNSNames`, standardizing `identifier.FromDNS` as the singular and `identifiers.FromDNS` as the plural (i.e. slice).

Add a named type `ACMEIdentifiers` so that we can add methods to slices. We will have a lot of slice handling code coming up, which this will make more elegant and readable.

Part of #7311