identifier: Rename FromDNSNames & AsProto; add ACMEIdentifiers named type
Rename `FromDNSNames` to `NewDNSSlice`, since it's exactly `NewDNS` except for slices.

Rename `AsProto` to use the "To" prefix, since it's the opposite of "From".

Add a named type `ACMEIdentifiers` so that we can add methods to slices. We will have a lot of slice handling code coming up, which this will make more elegant and readable.

Add a comment to explain naming conventions in the `identiifer` package.

Part of #7311
Alternative to #8068