Fix profile comparison when looking for authzs to reuse (#8144)
Previously, if the request asked for a profile, we were comparing the
address of that requested profile to the address of the profile field of
the found authz. Obviously these addresses were never the same. Instead,
compare the actual values, with an added nil check for safety.

This fixes a bug reported on the community forum. The updated test fails
without the accompanying code change.