Give registrations.contact column a default value (#8207)
Alter the "registrations" table so that the "contact" column has a
default value of the JSON empty list "[]". This, once deployed to all
production environments, will allow Boulder to stop writing to and
reading from this column, in turn allowing it to be eventually wholly
dropped from the database.

IN-11365 tracks the corresponding production database changes
Part of https://github.com/letsencrypt/boulder/issues/8176