Add AddressesResolved to tls-alpn-01 IP validation records (#8288)
Set the ValidationRecord.AddressesResolved field when performing
tls-alpn-01 validation on an IP address. This field is redundant with
ValidationRecord.Hostname, because we don't perform any resolution of
DNS names to IP addresses when conducting validation of an IP address.
But populating it anyway appeases the safety checks within the
Challenge.RecordsSane() method, and matches the fact that we populate it
for http-01 validation of IP addresses.

Add a check to several unit tests which fails on main but passes with
this change applied.

Fixes https://github.com/letsencrypt/boulder/issues/8286