Build Boulder in a container for release (#8331)
This creates a Dockerfile that can be used to build Boulder in a
container. It produces a minimal image that contains just the current
contents of our .deb and .tar.gz release artifacts: Boulder plus some
ancillary files.

The .deb and .tar.gz release artifacts are now produced downstream of
that minimal image, by extracting the file from it in the appropriate
format.

This removes the use of make-assets.sh in favor of container-build.sh. The
part of make-assets.sh that builds a .deb is extracted into make-deb.sh.
That script runs inside the minimal built Boulder image and outputs a
.deb to a volume that corresponds to the Boulder directory.

Fixes #8327

---------

Co-authored-by: James Renken <james@renken.is>