sfe/test: Configure SFE to use a zendesk-test-srv (#8354)
**test/zendeskfake:**
Address some optional comments from
https://github.com/letsencrypt/boulder/pull/8320.

**sfe/zendesk:**
Address some optional comments from
https://github.com/letsencrypt/boulder/pull/8320.

**ratelimits/overriderequests**:
- Remove unused constants that are not yet necessary.
- Remove Fundraising from tickets, we'll handle this value at submission
time.
- Comment the various Create*OverrideTicket functions.
- Pass the correct rate limit for the ticket being created.

**test/zendesk-test-srv:**
Testing the user experience of the entire submission flow will require a
non-nil zendeskClient in the SFE. Add a new test-only Zendesk server
implementation that allows us to configure and use the Zendesk client in
the SFE.

**sfe/templates**:
In a future PR, new Content Security Policy headers necessary to secure
our override request forms will disallow inline CSS. To keep that PR
focused (and the CSS diff easier to parse), this change moves the
existing CSS to a static file and references it in the existing HTML
template.

**sfe/sfe:**
Configure and assign the zendeskClient in config-next.

**cmd/sfe/main.go:**
Change AccountID configuration field to match AccountURI present in the
form.