Remove "ocsp-response" Ceremony type (#8396)
Remove the configuration, validation, logic, and tests which support the
"ocsp-response" ceremony type. We have not used this ceremony type to
produce OCSP responses in many years, and do not anticipate doing so at
any time in the foreseeable future. None of our CA certificates contain
AIA OCSP URIs, so we could not serve such responses even if we did
generate them.