Ceremony: remove support for AIA OCSP URIs (#8397)
Remove the ability to configure an ocsp-url from the ceremony tool's
root, intermediate, cross-sign, and csr ceremony types. We have not used
this field in our last several ceremonies, and do not plan to go back to
including AIA OCSP URIs in our CA certificates at any time in the
foreseeable future.