Remove support for temporally-sharded CRLs (#8400)
In crl-updater, delete the code which computes the set of temporal
chunks that belong to a shard, which queries the database for revoked
certs falling into those chunks, and which combines and deduplicates
those results against the revoked certs from explicit shards.

In the SA, make the ShardIdx argument to RevokeCertificate and
UpdateRevokedCertificate fully mandatory, and make updating the
revokedCertificates table unconditional.

This paves the way for deleting the sa.GetMaxExpiration and
sa.GetRevokedCerts methods, which were only called from code deleted by
this change.

IN-11747 tracks the corresponding config change in prod.

Part of https://github.com/letsencrypt/boulder/issues/8399
Part of https://github.com/letsencrypt/boulder/issues/8322