Use GetRevocationStatus instead of GetCertificateStatus (#8402)
In the SA, change the implementation of GetRevocationStatus to read from
the revokedCertificates table instead of reading from the
certificateStatus table.

In the WFE, switch to calling GetRevocationStatus when computing ARI
windows. Similarly, in the RA, make the same switch when checking if a
to-be-revoked certificate is already revoked.

Across all three locations, use new core constants to represent "good"
and "revoked", to avoid references to OCSP and unwieldy string/int
conversions.

This paves the way for removing sa.GetCertificateStatus, which now has
only one remaining caller which is not quite so easily changed.

Part of https://github.com/letsencrypt/boulder/issues/8322

(cherry picked from commit a5cf3725098fc487116173170fb1f55d2117a1af)