Start accepting a new log checksum format
This adds a new log checksum format without the variable-width integer encoding
currently used in Boulder. It's still CRC32, just encoded directly into base64
instead of with an extra layer of varint-encoding. Note that despite using the
varint encoding, Boulder always writes the full buffer out, so it's zero-padded
to 7 bytes.

This doesn't switch to using the new format yet, but just accepts it in the log
validator.

This is change 1/3 in a sequence:

* Allow new format
* Switch to new format
* Remove support for old format.

Each of these 3 changes should land one release apart.