Revert "Use GetRevocationStatus instead of GetCertificateStatus" (#8426)
This reverts https://github.com/letsencrypt/boulder/pull/8402

The revokedCertificates table does not have an index on the `serial`
column, unlike the certificateStatus table. This means that these
highly-targeted, single-certificate lookups are actually very
inefficient, and led to performance problems in the database.

This revert will be followed by a schema change to add an index to the
revokedCertificates table, and then by a reland of the original PR.

Part of https://github.com/letsencrypt/boulder/issues/8322