Add an index on serials to the revokedCertificates table (#8427)
Give the `revokedCertificates` table a non-unique index on the `serial`
column. This exactly matches the existing index on the
`certificateStatus` table.

We did not include this index initially because we were optimizing this
table for the crl-updater's crawl-by-shard-and-expiration behavior, but
this index is necessary to look up if a certificate has already been
revoked when computing ARI windows and when processing revocation
requests.

Part of https://github.com/letsencrypt/boulder/issues/8322