Refactor CA unittests to focus on top-level IssueCertificate (#8422)
Rewrite the vast majority of the CA unittests so that they test the
exported functions (NewCertificateAuthorityImpl and IssueCertificate)
rather than the non-exported helper functions (issuePrecertificate and
issueCertificateForPrecertificate). This makes the tests more robust to
future refactoring of the CA code itself, and will increase confidence
in the correctness of those refactors.

Part of https://github.com/letsencrypt/boulder/issues/8390