Remove requirement for IncludeCRLDistributionPoints config
The IncludeCRLDistributionPoints config item is ignored -- we include a CRLDP no matter what. Therefore the stanza requiring that one be set in the config is redundant and useless, and prevents us from cleaning up that config item.

It is likely that, if and when we are able to drop the CRLDP from short-lived certs, we will use a new "OmitRevocation" config field.