bdns: remove logDNSError (#8445)
This had some special casing for the "id mismatch" error we were
tracking for a while. We've resolved that issue by switching to DoH and
now see 0 instances of "id mismatch." However, it's still been
occasionally useful to see the detailed DNS errors, so I'm keeping a
simplified log line.