Refactor getOne out of GetSCTs
This is a small refactor to remove a pair of closures so we can directly launch
the submission goroutines with `go ctp.getOne(...)`.

This is itself pulled out of a slightly larger refactor which pulls the stagger
sleep back out of this function, but it's much cleaner to do this refactor as
a preliminary change.

There is a small functional change here: The same subCtx is passed to the
submission RPC. I think that shouldn't cause any problems.