Refactor GetSCTs into a central loop
Instead of spawning all goroutines immediately with their own sleeps, this uses
a ticker on the stagger interval to start additional submissions.

This shouldn't change much, but unblocks future improvements like kicking off
additional submissions immediately on failure to submit to a log, or adding
more logic to what order we submit in based on the current results.

As a small perf win, this should slightly decrease the number of goroutines as
we don't typically have to submit to too many logs.