Export prometheus client_golang's Go build_info metric (#8482)
We have our own version metric already, but supporting client_golang's
build_info means we can be more compatible with other Golang programs we
run which export this metric.

We can consider deprecating and removing our own metric in the future,
but we have dashboards and alerts using it, so there's no rush to do
that.