Remove dead code (#8506)
This started with noticing the semaphore package (forked for use in
OCSP) was dead.

That followed by using the golang `deadcode` tool and Goland's various
"Unused ..." inspection tools to find other dead code.

Nothing in here looks like it would be used externally to Boulder, as
far as I can tell.