Store authzIDs directly in order table (#8460)
The `orders` table gets a new column `authzs`, containing a
protobuf-encoded list of authorization IDs as int64s. Storing this data
directly in the `orders` table will allow us to get rid of the
`orderToAuthz2` table, which is large (#8451).

There's a new feature flag, `StoreAuthzsInOrders`, which controls
whether the SA expects the new column to exist. The SA gracefully
handles the case where the new column is NULL, by querying the
`orderToAuthz2` table. Eventually all valid orders will have a non-NULL
`authzs` column and we can remove the fallback.

I removed `orderToModel` because it had only one call site, and that
call site only used a subset of fields. This avoids having to update
`orderToModel` to know about encoding authzs, only to have that code
actually go unused in practice.