Remove usage of subtle.ConstantTimeCompare in validation (#8519)
While constant-time comparison is important in cryptographic algorithms,
that's not what we're doing here. The validation random token is not
intended to be secret in the same way as (say) a private key is: it's
just meant to be random enough that it's unlikely to exist in DNS or on
a webserver by chance. Possession of the random token does not give an
attacker any advantages; they still need to control the domain itself,
at which point they could get a random token of their own. Using
subtle.ConstantTimeCompare is overkill and sets a bad example for places
that truly do need to use it.