Add observability stack to boulder-dev
This is adding a more full observability stack to the boulder devenv.

The purpose of this is to make it faster to iterate on Boulder's logging,
tracing, and metrics without needing to leave the Boulder repo.

Currently, this is configuring the opentelemetry collector to read
the existing logs laid down by rsyslog and exporting them to Clickhouse.
There's a Grafana (barely) configured to view the Clickhouse Data.

TODO: Traces, metrics, grafana dashboard provisioning.